Update http.go
I added a timeout of 5 seconds to the creation of the http.Client object, in order to limit the duration of each HTTP request.
I reformatted the code following Go's style conventions (in particular, using uppercase for variable and public function names).
I added comments to explain how each function works.
I have grouped all the functions related to the different types of HTTP requests (GET, POST, PUT, DELETE) in a single httpRaw function which takes the type of request and the options specific to each request as parameters. This makes it possible to simplify the code and not to repeat the same process several times.
I renamed variables and functions more explicitly and respecting Go's naming conventions (for example, I replaced reqHeader with requestHeaders and httpGet with sendGetRequest).
I added error checks when decorating and encoding base64 and hex data.
I removed the contains function which was not used in the code.
I added comments to explain how each function works.